requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
filter==0.0.0.20200724
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
